requirements.txt
fastapi
pydantic
uvicorn
python-multipart
supabase
google-genai
httpx
sse-starlette
python-dotenv
anthropic
trafilatura
pyyaml
pypdf
python-docx
